# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Qillin AI Proxy API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MessageResponse {
  message: string;
}

export interface RegisterInput {
  email: string;
  /** @minLength 8 */
  password: string;
  name: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  user: 'user',
  admin: 'admin',
} as const;

export interface User {
  id: string;
  email: string;
  name: string;
  role: UserRole;
  qredits: number;
  isActive: boolean;
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export interface ModelUsage {
  modelId: string;
  modelName: string;
  requests: number;
  tokensIn: number;
  tokensOut: number;
  spend: number;
}

export interface RequestRecord {
  id: string;
  model: string;
  tokensIn: number;
  tokensOut: number;
  spend: number;
  timestamp: string;
}

export interface UsageStats {
  totalSpend: number;
  totalRequests: number;
  modelBreakdown: ModelUsage[];
  recentRequests: RequestRecord[];
}

export interface ApiKey {
  keyHash: string;
  name: string;
  /** @nullable */
  token: string | null;
  createdAt: string;
  /** @nullable */
  lastUsed: string | null;
  spend: number;
}

export interface ApiKeyInput {
  name: string;
  /** @nullable */
  maxBudget?: number | null;
}

export interface Model {
  id: string;
  name: string;
  litellmModel: string;
  provider: string;
  providerId: string;
  contextWindow: number;
  inputCostPerMtok: number;
  outputCostPerMtok: number;
  enabled: boolean;
}

export interface ModelInput {
  name: string;
  litellmModel: string;
  providerId: string;
  contextWindow: number;
  inputCostPerMtok: number;
  outputCostPerMtok: number;
  enabled?: boolean;
}

export interface DailySpend {
  date: string;
  spend: number;
  requests: number;
}

export interface UserSpendSummary {
  userId: string;
  email: string;
  name: string;
  spend: number;
  requests: number;
}

export interface AdminStats {
  totalUsers: number;
  totalSpend: number;
  totalRequests: number;
  activeModels: number;
  activeProviders: number;
  spendByDay: DailySpend[];
  topUsers: UserSpendSummary[];
}

export interface ActivityRecord {
  id: string;
  type: string;
  /** @nullable */
  userId?: string | null;
  /** @nullable */
  userEmail?: string | null;
  /** @nullable */
  model?: string | null;
  /** @nullable */
  tokensIn?: number | null;
  /** @nullable */
  tokensOut?: number | null;
  /** @nullable */
  spend?: number | null;
  timestamp: string;
}

export type AdminUserRole = typeof AdminUserRole[keyof typeof AdminUserRole];


export const AdminUserRole = {
  user: 'user',
  admin: 'admin',
} as const;

export interface AdminUser {
  id: string;
  email: string;
  name: string;
  role: AdminUserRole;
  qredits: number;
  totalSpend: number;
  totalRequests: number;
  /** @nullable */
  allowedModels?: string[] | null;
  isActive: boolean;
  createdAt: string;
  /** @nullable */
  litellmUserId?: string | null;
}

export interface UserUpdate {
  /** @nullable */
  qredits?: number | null;
  /** @nullable */
  isActive?: boolean | null;
  /** @nullable */
  allowedModels?: string[] | null;
  /** @nullable */
  role?: string | null;
}

export type CreditInputOperation = typeof CreditInputOperation[keyof typeof CreditInputOperation];


export const CreditInputOperation = {
  add: 'add',
  set: 'set',
  subtract: 'subtract',
} as const;

export interface CreditInput {
  amount: number;
  operation: CreditInputOperation;
}

export type ProviderType = typeof ProviderType[keyof typeof ProviderType];


export const ProviderType = {
  openai: 'openai',
  anthropic: 'anthropic',
  custom: 'custom',
} as const;

export type ProviderLoadBalancing = typeof ProviderLoadBalancing[keyof typeof ProviderLoadBalancing];


export const ProviderLoadBalancing = {
  round_robin: 'round_robin',
  priority: 'priority',
} as const;

export interface ProviderApiKey {
  id: string;
  /** @nullable */
  label?: string | null;
  keyMasked: string;
  priority: number;
  failCount: number;
}

export interface Provider {
  id: string;
  name: string;
  type: ProviderType;
  /** @nullable */
  baseUrl?: string | null;
  loadBalancing: ProviderLoadBalancing;
  apiKeys: ProviderApiKey[];
  modelCount: number;
  isActive: boolean;
  createdAt: string;
}

export interface ProviderApiKeyInput {
  key: string;
  /** @nullable */
  label?: string | null;
  priority: number;
}

export type ProviderInputType = typeof ProviderInputType[keyof typeof ProviderInputType];


export const ProviderInputType = {
  openai: 'openai',
  anthropic: 'anthropic',
  custom: 'custom',
} as const;

export type ProviderInputLoadBalancing = typeof ProviderInputLoadBalancing[keyof typeof ProviderInputLoadBalancing];


export const ProviderInputLoadBalancing = {
  round_robin: 'round_robin',
  priority: 'priority',
} as const;

export interface ProviderInput {
  name: string;
  type: ProviderInputType;
  /** @nullable */
  baseUrl?: string | null;
  loadBalancing: ProviderInputLoadBalancing;
  apiKeys: ProviderApiKeyInput[];
}

export interface FetchModelsInput {
  apiKey: string;
  /** @nullable */
  baseUrl?: string | null;
}

export interface FetchedModel {
  id: string;
  name: string;
  /** @nullable */
  contextWindow?: number | null;
}