# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Qillin AI Proxy API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /models/{modelId}:
    get:
      operationId: getModel
      tags: [models]
      summary: Get a specific model
      parameters:
        - name: modelId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Model"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # Admin
...
  /admin/users/{userId}:
    get:
      operationId: getAdminUser
      tags: [admin]
      summary: Get a specific user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
    patch:
      operationId: updateAdminUser
      tags: [admin]
      summary: Update a user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
  /admin/users/{userId}/credits:
    post:
      operationId: addUserCredits
      tags: [admin]
      summary: Add or set user Qredits
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreditInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
  /admin/providers:
    get:
      operationId: getAdminProviders
      tags: [admin]
      summary: List all providers
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Provider"
    post:
      operationId: createAdminProvider
      tags: [admin]
      summary: Add a provider
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProviderInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Provider"
  /admin/providers/{providerId}:
    put:
      operationId: updateAdminProvider
      tags: [admin]
      summary: Update a provider
      parameters:
        - name: providerId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProviderInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Provider"
    delete:
      operationId: deleteAdminProvider
      tags: [admin]
      summary: Delete a provider
      parameters:
        - name: providerId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /admin/providers/{providerId}/fetch-models:
    post:
      operationId: fetchProviderModels
      tags: [admin]
      summary: Fetch available models from a provider using an API key
      parameters:
        - name: providerId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FetchModelsInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FetchedModel"
  /admin/models:
    get:
      operationId: getAdminModels
      tags: [admin]
      summary: List all models (admin view)
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Model"
    post:
      operationId: createAdminModel
      tags: [admin]
      summary: Add a model to the catalog
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ModelInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Model"
  /admin/models/{modelId}:
    put:
      operationId: updateAdminModel
      tags: [admin]
      summary: Update a model
      parameters:
        - name: modelId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ModelInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Model"
    delete:
      operationId: deleteAdminModel
      tags: [admin]
      summary: Delete a model
      parameters:
        - name: modelId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]

    # Auth
...
    Model:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        litellmModel:
          type: string
        provider:
          type: string
        providerId:
          type: string
        contextWindow:
          type: integer
        inputCostPerMtok:
          type: number
        outputCostPerMtok:
          type: number
        enabled:
          type: boolean
      required: [id, name, litellmModel, provider, providerId, contextWindow, inputCostPerMtok, outputCostPerMtok, enabled]
    ModelInput:
      type: object
      properties:
        name:
          type: string
        litellmModel:
          type: string
        providerId:
          type: string
        contextWindow:
          type: integer
        inputCostPerMtok:
          type: number
        outputCostPerMtok:
          type: number
        enabled:
          type: boolean
      required: [name, litellmModel, providerId, contextWindow, inputCostPerMtok, outputCostPerMtok]

    # Admin
...
    AdminUser:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [user, admin]
        qredits:
          type: number
        totalSpend:
          type: number
        totalRequests:
          type: integer
        allowedModels:
          type: ["array", "null"]
          items:
            type: string
        isActive:
          type: boolean
        createdAt:
          type: string
        litellmUserId:
          type: ["string", "null"]
      required: [id, email, name, role, qredits, totalSpend, totalRequests, isActive, createdAt]
    UserUpdate:
      type: object
      properties:
        qredits:
          type: ["number", "null"]
        isActive:
          type: ["boolean", "null"]
        allowedModels:
          type: ["array", "null"]
          items:
            type: string
        role:
          type: ["string", "null"]
    CreditInput:
      type: object
      properties:
        amount:
          type: number
        operation:
          type: string
          enum: [add, set, subtract]
      required: [amount, operation]
    Provider:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        type:
          type: string
          enum: [openai, anthropic, custom]
        baseUrl:
          type: ["string", "null"]
        loadBalancing:
          type: string
          enum: [round_robin, priority]
        apiKeys:
          type: array
          items:
            $ref: "#/components/schemas/ProviderApiKey"
        modelCount:
          type: integer
        isActive:
          type: boolean
        createdAt:
          type: string
      required: [id, name, type, loadBalancing, apiKeys, modelCount, isActive, createdAt]
    ProviderApiKey:
      type: object
      properties:
        id:
          type: string
        label:
          type: ["string", "null"]
        keyMasked:
          type: string
        priority:
          type: integer
        failCount:
          type: integer
      required: [id, keyMasked, priority, failCount]
    ProviderApiKeyInput:
      type: object
      properties:
        key:
          type: string
        label:
          type: ["string", "null"]
        priority:
          type: integer
      required: [key, priority]
    ProviderInput:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
          enum: [openai, anthropic, custom]
        baseUrl:
          type: ["string", "null"]
        loadBalancing:
          type: string
          enum: [round_robin, priority]
        apiKeys:
          type: array
          items:
            $ref: "#/components/schemas/ProviderApiKeyInput"
      required: [name, type, loadBalancing, apiKeys]
    FetchModelsInput:
      type: object
      properties:
        apiKey:
          type: string
        baseUrl:
          type: ["string", "null"]
      required: [apiKey]
    FetchedModel:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        contextWindow:
          type: ["integer", "null"]
      required: [id, name]