# lib/api-zod/src/generated/types/creditInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Qillin AI Proxy API
 * OpenAPI spec version: 0.1.0
 */
import type { CreditInputOperation } from './creditInputOperation';

export interface CreditInput {
  amount: number;
  operation: CreditInputOperation;
}