# lib/api-zod/src/generated/types/fetchedModel.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Qillin AI Proxy API
 * OpenAPI spec version: 0.1.0
 */

export interface FetchedModel {
  id: string;
  name: string;
  /** @nullable */
  contextWindow?: number | null;
}