# lib/api-zod/src/generated/types/modelInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Qillin AI Proxy API
 * OpenAPI spec version: 0.1.0
 */

export interface ModelInput {
  name: string;
  litellmModel: string;
  providerId: string;
  contextWindow: number;
  inputCostPerMtok: number;
  outputCostPerMtok: number;
  enabled?: boolean;
}