# lib/api-zod/src/generated/types/providerApiKeyInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Qillin AI Proxy API
 * OpenAPI spec version: 0.1.0
 */

export interface ProviderApiKeyInput {
  key: string;
  /** @nullable */
  label?: string | null;
  priority: number;
}