# lib/api-zod/src/generated/types/userUpdate.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Qillin AI Proxy API
 * OpenAPI spec version: 0.1.0
 */

export interface UserUpdate {
  /** @nullable */
  qredits?: number | null;
  /** @nullable */
  isActive?: boolean | null;
  /** @nullable */
  allowedModels?: string[] | null;
  /** @nullable */
  role?: string | null;
}